Finished

+++ exercises/conversions/using_as.rs
@@ -10,11 +10,14 @@
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
+// Type casting in Rust is done via the usage of the `as` operator.
+// Note that the `as` operator is not only used when type casting. It also helps
+// with renaming imports.
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    // TODO: Make a conversion before dividing.
+    total / values.len() as f64
 }
 
 fn main() {
